fix : check if HighFive can be potentially installed

+++ detail/build_HighFive.sh
@@ -26,23 +26,27 @@ function elastica_detect_compiler() {
 elastica_detect_compiler
 _CXX_COMPILER=${2:-"${_CXX_}"}
 
-# Requires >3.14
-cmake -B "${HIGHFIVE_BUILD_DIR}" \
-	-D CMAKE_BUILD_TYPE=Release \
-	-D CMAKE_CXX_COMPILER="${_CXX_COMPILER}" \
-	-D HIGHFIVE_USE_BOOST=OFF \
-	-D HIGHFIVE_USE_HALF_FLOAT=OFF \
-	-D HIGHFIVE_USE_EIGEN=OFF \
-	-D HIGHFIVE_USE_OPENCV=OFF \
-	-D HIGHFIVE_USE_XTENSOR=OFF \
-	-D HIGHFIVE_EXAMPLES=OFF \
-	-D HIGHFIVE_PARALLEL_HDF5=OFF \
-	-D HIGHFIVE_BUILD_DOCS=OFF \
-	-S .
-# doesn't need to be parallel as its header only
-cmake --build "${HIGHFIVE_BUILD_DIR}"
-cmake --install "${HIGHFIVE_BUILD_DIR}" --prefix "${HIGHFIVE_INSTALL_PREFIX}"
-
+# We need high-five to be found, else this steps error out. Rather than searching for the libraries which is harder,
+# we search for the compiler-wrapper here.
+if command -v h5cc >/dev/null 2>&1; then # h5cc exists, so good chance that Cmake wont bug out
+	cmake -B "${HIGHFIVE_BUILD_DIR}" \
+		-D CMAKE_BUILD_TYPE=Release \
+		-D CMAKE_CXX_COMPILER="${_CXX_COMPILER}" \
+		-D HIGHFIVE_USE_BOOST=OFF \
+		-D HIGHFIVE_USE_HALF_FLOAT=OFF \
+		-D HIGHFIVE_USE_EIGEN=OFF \
+		-D HIGHFIVE_USE_OPENCV=OFF \
+		-D HIGHFIVE_USE_XTENSOR=OFF \
+		-D HIGHFIVE_EXAMPLES=OFF \
+		-D HIGHFIVE_PARALLEL_HDF5=OFF \
+		-D HIGHFIVE_BUILD_DOCS=OFF \
+		-S .
+	# doesn't need to be parallel as its header only
+	cmake --build "${HIGHFIVE_BUILD_DIR}"
+	cmake --install "${HIGHFIVE_BUILD_DIR}" --prefix "${HIGHFIVE_INSTALL_PREFIX}"
+else
+	echo "h5cc was not found. Chances are that HDF5 is not installed on your system, so the HighFive dependency will not be installed."
+fi
 unset _CXX_COMPILER
 unset _CXX_
 unset -f elastica_detect_compiler
